Don't call pbr.version
Possible workaround to:
```
Exception: Versioning for this project requires either an sdist tarball, or access to an upstream git repository. It's also possible that there is a mismatch between the package name in setup.cfg and the argument given to pbr.version.VersionInfo. Project name sphinxcontrib-apidoc was given, but was not able to be found.
```